+++ src/routes/+layout.svelte
@@ -9,6 +9,7 @@
 	import CardEditorModal from '$lib/components/CardEditorModal.svelte';
 	import SleepTimer from '$lib/components/SleepTimer.svelte';
 	import MiniPlayer from '$lib/components/MiniPlayer.svelte';
+	import StreamsTray from '$lib/components/StreamsTray.svelte';
 	import DiscordPresence from '$lib/components/DiscordPresence.svelte';
 	import TitleInfoModal from '$lib/components/TitleInfoModal.svelte';
 	import HiddenTitlesModal from '$lib/components/HiddenTitlesModal.svelte';
@@ -108,6 +109,7 @@
 <CardEditorModal />
 <SleepTimer />
 <MiniPlayer />
+<StreamsTray />
 <DiscordPresence />
 <TitleInfoModal />
 <HiddenTitlesModal />